backend/requirements.txt
canvasapi==3.2.0
python-dotenv==1.0.1
Flask==3.0.2
Flask-CORS==4.0.0
gunicorn==22.0.0
firebase-admin==6.2.0
